src/main/java/me/namuhuchutong/step4/Car.java
package me.namuhuchutong.step4;

public class Car {

    private int position;
    private final String name;

    public Car(String name) {
        this(0, name);
    }

    public Car(int position, String name) {
        this.position = position;
        this.name = name;
    }

    void move(int value) {
        if (3 < value) {
            this.position += 1;
        }
    }

    public int getPosition() {
        return position;
    }

    public String getName() {
        return name;
    }
}
